tender_project/requirements.txt
# requirements.txt
Django==4.2.7  # (стабильная версия, совместимая с DRF)
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openpyxl==3.1.2  # для парсинга .xlsx
pandas==2.1.3   # для удобной работы с excel/csv (опционально, но сильно упростит жизнь)
django-cors-headers==4.3.1  # для связи с фронтендом
Pillow==10.1.0  # для работы с файлами/изображениями (если понадобятся)pip install -r requirements.txt